+++ src/shared/chat/message.js
@@ -116,7 +116,8 @@ export default class Message extends ObservableElement {
     /** @param {MouseEvent} ev */
     onImgClick(ev) {
         ev.preventDefault();
-        api.modal.show('converse-image-modal', { src: /** @type {HTMLImageElement} */(ev.target).src }, ev);
+        const img = /** @type {HTMLImageElement} */ (ev.target);
+        api.modal.show('converse-image-modal', { src: img.src, filename: img.dataset.filename }, ev);
     }
 
     onUnfurlAnimationEnd() {