+++ src/shared/modals/image.js
@@ -3,7 +3,7 @@ import tplImageModal from "./templates/image.js";
 import { __ } from 'i18n';
 import { api } from "@converse/headless";
 import { getFileName } from 'utils/html.js';
-import { html } from "lit";
+import { html, nothing } from "lit";
 
 import './styles/image.scss';
 
@@ -16,6 +16,9 @@ export default class ImageModal extends BaseModal {
     constructor (options) {
         super(options);
         this.src = options.src;
+        // The original filename, used as the `download` name when `src` is an
+        // opaque `blob:` URL (e.g. a decrypted OMEMO image). See #2632.
+        this.filename = options.filename;
     }
 
     renderModal () {
@@ -23,7 +26,14 @@ export default class ImageModal extends BaseModal {
     }
 
     getModalTitle () {
-        return html`${__('Image: ')}<a target="_blank" rel="noopener" href="${this.src}">${getFileName(this.src)}</a>`;
+        const filename = this.filename || getFileName(this.src);
+        return html`${__('Image: ')}<a
+                target="_blank"
+                rel="noopener"
+                download="${this.filename || nothing}"
+                href="${this.src}"
+                >${filename}</a
+            >`;
     }
 }