+++ src/shared/texture/directives/image.js
@@ -1,4 +1,4 @@
-import { html } from "lit";
+import { html, nothing } from "lit";
 import { AsyncDirective } from "lit/async-directive.js";
 import { directive } from "lit/directive.js";
 import { u } from "@converse/headless";
@@ -12,14 +12,21 @@ class ImageDirective extends AsyncDirective {
      * @param {string} [href] - The optional hyperlink for the image.
      * @param {Function} [onLoad] - Callback function to be called once the image has loaded.
      * @param {Function} [onClick] - Callback function to be called once the image has been clicked.
+     * @param {string} [filename] - The original filename, used as the download name when
+     *  the `src` is an opaque URL (e.g. a `blob:` URL for a decrypted OMEMO image).
      * @returns {import('lit').TemplateResult}
      */
-    render(src, href, onLoad, onClick) {
+    render(src, href, onLoad, onClick, filename) {
         return href
-            ? html`<a href="${href}" class="chat-image__link" target="_blank" rel="noopener"
-                  >${this.renderImage(src, href, onLoad, onClick)}</a
+            ? html`<a
+                  href="${href}"
+                  class="chat-image__link"
+                  target="_blank"
+                  rel="noopener"
+                  download="${filename || nothing}"
+                  >${this.renderImage(src, href, onLoad, onClick, filename)}</a
               >`
-            : this.renderImage(src, href, onLoad, onClick);
+            : this.renderImage(src, href, onLoad, onClick, filename);
     }
 
     /**
@@ -27,14 +34,16 @@ class ImageDirective extends AsyncDirective {
      * @param {string} [href] - The optional hyperlink for the image.
      * @param {Function} [onLoad] - Callback function to be called once the image has loaded.
      * @param {Function} [onClick] - Callback function to be called once the image has been clicked.
+     * @param {string} [filename] - The original filename of the image.
      * @returns {import('lit').TemplateResult}
      */
-    renderImage(src, href, onLoad, onClick) {
+    renderImage(src, href, onLoad, onClick, filename) {
         return html`<img class="chat-image img-thumbnail"
                     loading="lazy"
                     src="${src}"
+                    data-filename="${filename || nothing}"
                     @click=${onClick}
-                    @error=${() => this.onError(src, href, onLoad, onClick)}
+                    @error=${() => this.onError(src, href, onLoad, onClick, filename)}
                     @load="${onLoad}"/></a>`;
     }
 
@@ -44,8 +53,9 @@ class ImageDirective extends AsyncDirective {
      * @param {string} [href] - The optional hyperlink for the image.
      * @param {Function} [onLoad] - Callback function to be called once the image has loaded.
      * @param {Function} [onClick] - Callback function to be called once the image has been clicked.
+     * @param {string} [filename] - The original filename of the image.
      */
-    onError(src, href, onLoad, onClick) {
+    onError(src, href, onLoad, onClick, filename) {
         if (!this.isConnected) {
             return href ? getHyperlinkTemplate(href) : html`<span>Image failed to load</span>`;
         }
@@ -55,11 +65,11 @@ class ImageDirective extends AsyncDirective {
         } else {
             try {
                 const url = new URL(src);
-                const filename = url.pathname.split("/").pop();
-                if (filename) {
-                    const new_filename = `${filename}.png`;
-                    url.pathname = url.pathname.replace(filename, new_filename);
-                    this.setValue(renderImage(url.toString(), href, onLoad, onClick));
+                const fallback_filename = url.pathname.split("/").pop();
+                if (fallback_filename) {
+                    const new_filename = `${fallback_filename}.png`;
+                    url.pathname = url.pathname.replace(fallback_filename, new_filename);
+                    this.setValue(renderImage(url.toString(), href, onLoad, onClick, filename));
                 }
             } catch (error) {
                 console.error("Invalid URL:", src);
@@ -77,5 +87,6 @@ class ImageDirective extends AsyncDirective {
  * @param { String } href - The value that will be assigned to the `href` attribute of the `<img>` element.
  * @param { Function } onLoad - A callback function to be called once the image has loaded.
  * @param { Function } onClick - A callback function to be called once the image has been clicked.
+ * @param { String } filename - The original filename, used as the `download` name for opaque URLs.
  */
 export const renderImage = directive(ImageDirective);